new ecmean library
